+++ benchmark/StorageBenchmark.kt
@@ -14,27 +14,25 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package simulation
+package com.vaticle.typedb.benchmarks.storage
 
-import com.vaticle.typedb.benchmarks.storage.AgentFactory
-import com.vaticle.typedb.benchmarks.storage.PersonAgent
 import com.vaticle.typedb.client.api.TypeDBOptions
 import com.vaticle.typedb.client.api.TypeDBSession
 import com.vaticle.typedb.benchmarks.storage.common.Context
-import com.vaticle.typedb.simulation.common.seed.RandomSource
-import com.vaticle.typedb.simulation.typedb.TypeDBClient
+import com.vaticle.typedb.benchmark.framework.common.seed.RandomSource
+import com.vaticle.typedb.benchmark.framework.typedb.TypeDBClient
 import mu.KotlinLogging
 import java.nio.file.Paths
 
 class StorageBenchmark internal constructor(client: TypeDBClient, context: Context) :
-    com.vaticle.typedb.simulation.typedb.TypeDBSimulation<Context>(client, context, AgentFactory(client, context)) {
+    com.vaticle.typedb.benchmark.framework.typedb.TypeDBSimulation<Context>(client, context, AgentFactory(client, context)) {
 
     private val LOGGER = KotlinLogging.logger {}
     override val agentPackage = PersonAgent::class.java.packageName
     override val name = "StorageBenchmark"
 
     // TODO: Update this filepath
-    override val schemaFiles = listOf(Paths.get("schema.tql").toFile())
+    override val schemaFiles = listOf(Paths.get("benchmark/schema.tql").toFile())
     private val options = TypeDBOptions.core().infer(true)
 
     override fun initData(nativeSession: TypeDBSession, randomSource: RandomSource) {
