+++ benchmark/common/BUILD
@@ -25,8 +25,8 @@ kt_jvm_library(
     deps = [
         # External Vaticle Dependencies
         "@vaticle_typedb_common//:common",
-        "@vaticle_typedb_simulation//simulation/common",
-        "@vaticle_typedb_simulation//simulation:typedb_simulation",
+        "//framework/common",
+        "//framework:typedb_simulation",
         "@vaticle_typedb_client_java//common",
         "@vaticle_typedb_client_java//api",
