+++ benchmark/config/ci-tests.yml
@@ -1,0 +1,65 @@
+#
+# Copyright (C) 2022 Vaticle
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+#
+
+agents:
+
+  - name: "PersonAgent"
+    action: "createPerson"
+    runsPerIteration: 5
+    trace: false
+
+  - name: "PersonAgent"
+    action: "createFriendship"
+    runsPerIteration: 10
+    trace: false
+
+  - name: "PersonAgent"
+    action: "readAddressFromName"
+    runsPerIteration: 1
+    trace: false
+
+  - name: "PersonAgent"
+    action: "readFriendsOf"
+    runsPerIteration: 1
+    trace: false
+
+  - name: "PersonAgent"
+    action: "readFriendsOfFriends"
+    runsPerIteration: 1
+    trace: false
+
+  - name: "PersonAgent"
+    action: "readPersonsByPostCode"
+    runsPerIteration: 10
+    trace: false
+
+traceSampling:
+  # Options: `"every"` for every K traces; `"log"` for logarithm with base N
+  function: "every"
+  arg: 10
+
+run:
+  randomSeed: 1
+  iterations: 5
+  partitions: 4
+  databaseName: "benchmark-on-ci"
+  recreateDatabase: true
+
+model:
+  personPerIteration: 10
+  friendshipPerIteration: 50
+  postCodes: 97
