+++ framework/typedb/TypeDBSessionEx.kt
@@ -14,13 +14,14 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.vaticle.typedb.simulation.neo4j
+package com.vaticle.typedb.benchmark.framework.typedb
 
-import org.neo4j.driver.Session
+import com.vaticle.typedb.client.api.TypeDBOptions
+import com.vaticle.typedb.client.api.TypeDBSession
+import com.vaticle.typedb.client.api.TypeDBTransaction.Type.READ
+import com.vaticle.typedb.client.api.TypeDBTransaction.Type.WRITE
 
-object Neo4jSessionEx {
-
-//    fun Session.readTransaction() {
-//
-//    }
+object TypeDBSessionEx {
+    fun TypeDBSession.readTransaction(infer: Boolean) = transaction(READ, TypeDBOptions.core().infer(infer))
+    fun TypeDBSession.writeTransaction() = transaction(WRITE)
 }
